+++ figma/code.ts
@@ -1,3 +1,4 @@
+/// <reference types="@figma/plugin-typings" />
 figma.showUI(__html__, {
   width: 420,
   height: 220
