+++ src/components/DatetimeString.vue
@@ -1,0 +1,8 @@
+<script setup lang="ts">
+</script>
+
+<template>
+  <div class="datetime-string">
+    Date
+  </div>
+</template>
