--- src/components/HelloWorld.vue
@@ -1,13 +1,0 @@
-<script setup lang="ts">
-defineProps<{ msg: string }>()
-</script>
-
-<template>
-  <p>{{ msg }}</p>
-</template>
-
-<style scoped>
-.read-the-docs {
-  color: #888;
-}
-</style>
